# backend/openapi_dump.json
{"openapi":"3.1.0","info":{"title":"AgentStudioBackend","description":"FastAPI Backend for Agent Studio with Global Tools, MCP Client, and OpenAI Integration","version":"1.0.0"},"paths":{"/api/v1/tools":{"get":{"tags":["Tools Registry"],"summary":"List Tools","description":"Mengambil daftar seluruh tools yang terdaftar pada registry allowlist.","operationId":"list_tools_api_v1_tools_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ToolResponse"},"type":"array","title":"Response List Tools Api V1 Tools Get"}}}}},"security":[{"APIKeyHeader":[]}]},"post":{"tags":["Tools Registry"],"summary":"Register Tool","description":"Mendaftarkan custom/built-in tool ke registry catalog.","operationId":"register_tool_api_v1_tools_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ToolRegisterRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ToolResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyHeader":[]}]}},"/api/v1/mcp":{"get":{"tags":["MCP Servers"],"summary":"List Mcp Servers","description":"Mengambil daftar seluruh MCP Servers yang terdaftar.","operationId":"list_mcp_servers_api_v1_mcp_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/MCPServerResponse"},"type":"array","title":"Response List Mcp Servers Api V1 Mcp Get"}}}}},"security":[{"APIKeyHeader":[]}]},"post":{"tags":["MCP Servers"],"summary":"Register Mcp Server","description":"Mendaftarkan MCP Server baru, menghubungkan MCP Client, dan otomatis mengimpor tools.","operationId":"register_mcp_server_api_v1_mcp_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MCPServerRegisterRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MCPServerResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyHeader":[]}]}},"/api/v1/builder/chat":{"post":{"tags":["Agent Builder"],"summary":"Builder Chat","description":"Endpoint untuk menyusun spesifikasi Agent berbasis bahasa alami dari pengguna.","operationId":"builder_chat_api_v1_builder_chat_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BuilderChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BuilderChatResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyHeader":[]}]}},"/api/v1/agent/chat":{"post":{"tags":["Agent Execution"],"summary":"Agent Chat","description":"Endpoint eksekusi percakapan agent secara synchronous.","operationId":"agent_chat_api_v1_agent_chat_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AgentChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AgentChatResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyHeader":[]}]}},"/api/v1/agent/chat/stream":{"post":{"tags":["Agent Execution"],"summary":"Agent Chat Stream","description":"Endpoint eksekusi percakapan agent secara real-time streaming SSE.","operationId":"agent_chat_stream_api_v1_agent_chat_stream_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AgentChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyHeader":[]}]}},"/api/v1/agent/trace/{run_id}":{"get":{"tags":["Trace & Simulations"],"summary":"Get Trace Log","description":"Mengambil detail timeline eksekusi trace log berdasarkan run_id.","operationId":"get_trace_log_api_v1_agent_trace__run_id__get","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"run_id","in":"path","required":true,"schema":{"type":"string","title":"Run Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TraceRunResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/agent/simulations":{"get":{"tags":["Trace & Simulations"],"summary":"List Simulations","description":"Mengambil daftar test cases skenario pengujian Agent.","operationId":"list_simulations_api_v1_agent_simulations_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/SimulationCase"},"type":"array","title":"Response List Simulations Api V1 Agent Simulations Get"}}}}},"security":[{"APIKeyHeader":[]}]}},"/health":{"get":{"tags":["Health Check"],"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AgentChatRequest":{"properties":{"agent_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Agent Id"},"agent_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Agent Name","default":"BPJS Customer Service Agent"},"message":{"type":"string","title":"Message"},"stream":{"type":"boolean","title":"Stream","default":false}},"type":"object","required":["message"],"title":"AgentChatRequest"},"AgentChatResponse":{"properties":{"id":{"type":"string","title":"Id"},"sender":{"type":"string","title":"Sender","default":"agent"},"text":{"type":"string","title":"Text"},"toolCall":{"anyOf":[{"$ref":"#/components/schemas/ToolCallDetail"},{"type":"null"}]},"status":{"type":"string","title":"Status","default":"ok"},"run_id":{"type":"string","title":"Run Id"},"time":{"type":"string","title":"Time","default":"Sekarang"}},"type":"object","required":["id","text","run_id"],"title":"AgentChatResponse"},"AgentSpec":{"properties":{"name":{"type":"string","title":"Name"},"description":{"type":"string","title":"Description"},"instructions":{"type":"string","title":"Instructions"},"model":{"type":"string","title":"Model","default":"gpt-4o-mini"},"tools":{"items":{"type":"string"},"type":"array","title":"Tools","default":[]},"mcp_servers":{"items":{"type":"string"},"type":"array","title":"Mcp Servers","default":[]},"harness":{"type":"string","title":"Harness","default":"default-safe-v1"}},"type":"object","required":["name","description","instructions"],"title":"AgentSpec"},"BuilderChatRequest":{"properties":{"prompt":{"type":"string","title":"Prompt"},"current_spec":{"anyOf":[{"$ref":"#/components/schemas/AgentSpec"},{"type":"null"}]}},"type":"object","required":["prompt"],"title":"BuilderChatRequest"},"BuilderChatResponse":{"properties":{"id":{"type":"string","title":"Id"},"message":{"type":"string","title":"Message"},"spec":{"$ref":"#/components/schemas/AgentSpec"}},"type":"object","required":["id","message","spec"],"title":"BuilderChatResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MCPServerRegisterRequest":{"properties":{"name":{"type":"string","title":"Name"},"url":{"type":"string","title":"Url"}},"type":"object","required":["name","url"],"title":"MCPServerRegisterRequest"},"MCPServerResponse":{"properties":{"id":{"type":"string","title":"Id"},"name":{"type":"string","title":"Name"},"url":{"type":"string","title":"Url"},"status":{"type":"string","title":"Status"},"tools_count":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Tools Count","default":0},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","name","url","status","created_at"],"title":"MCPServerResponse"},"SimulationCase":{"properties":{"id":{"type":"string","title":"Id"},"scenario":{"type":"string","title":"Scenario"},"input":{"type":"string","title":"Input"},"expected":{"type":"string","title":"Expected"},"actual":{"type":"string","title":"Actual"},"status":{"type":"string","title":"Status","default":"passed"},"duration":{"type":"string","title":"Duration"}},"type":"object","required":["id","scenario","input","expected","actual","duration"],"title":"SimulationCase"},"ToolCallDetail":{"properties":{"name":{"type":"string","title":"Name"},"params":{"additionalProperties":true,"type":"object","title":"Params"},"result":{"title":"Result"},"duration_ms":{"type":"integer","title":"Duration Ms"}},"type":"object","required":["name","params","result","duration_ms"],"title":"ToolCallDetail"},"ToolRegisterRequest":{"properties":{"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"source_type":{"type":"string","title":"Source Type","default":"builtin"},"input_schema":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Input Schema"}},"type":"object","required":["name"],"title":"ToolRegisterRequest"},"ToolResponse":{"properties":{"id":{"type":"string","title":"Id"},"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"source":{"type":"string","title":"Source"},"auth":{"type":"string","title":"Auth"},"health":{"type":"string","title":"Health"},"usedBy":{"type":"string","title":"Usedby"},"input_schema":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Input Schema"},"mcp_server_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Mcp Server Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","name","source","auth","health","usedBy","created_at"],"title":"ToolResponse"},"TraceRunResponse":{"properties":{"run_id":{"type":"string","title":"Run Id"},"status":{"type":"string","title":"Status","default":"OK"},"total_duration_ms":{"type":"integer","title":"Total Duration Ms"},"total_tokens":{"type":"integer","title":"Total Tokens"},"steps":{"items":{"$ref":"#/components/schemas/TraceStepResponse"},"type":"array","title":"Steps"}},"type":"object","required":["run_id","total_duration_ms","total_tokens","steps"],"title":"TraceRunResponse"},"TraceStepResponse":{"properties":{"step_no":{"type":"integer","title":"Step No"},"type":{"type":"string","title":"Type"},"title":{"type":"string","title":"Title"},"duration_ms":{"type":"integer","title":"Duration Ms"},"detail":{"title":"Detail"},"status":{"type":"string","title":"Status","default":"ok"}},"type":"object","required":["step_no","type","title","duration_ms","detail"],"title":"TraceStepResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"APIKeyHeader":{"type":"apiKey","in":"header","name":"X-API-Key"}}}}